Add migration, user registration, user login

+++ App/src/main/Models/Product.cs
@@ -11,5 +11,5 @@ public class Product
     public int ProductId { get; set; }
     public string ProductName { get; set; }
     public string? ProductDescription { get; set; }
-    public float? ProductPrice { get; set; }
+    public decimal? ProductPrice { get; set; }
 }
